Decouple sidebar index creation logic
