+++ lib/server/readMetadata.js
@@ -12,6 +12,7 @@ const fs = require('fs');
 const glob = require('glob');
 
 const metadataUtils = require('./metadataUtils');
+const createDocsIndex = require('./metadata/createDocsIndex');
 
 const env = require('./env.js');
 const blog = require('./blog.js');
@@ -42,45 +43,20 @@ const SupportedHeaderFields = new Set([
 function getDocsPath() {
   return siteConfig.customDocsPath ? siteConfig.customDocsPath : 'docs';
 }
+
 // returns map from id to object containing sidebar ordering info
 function readSidebar() {
   let allSidebars;
+
   if (fs.existsSync(`${CWD}/sidebars.json`)) {
     allSidebars = require(`${CWD}/sidebars.json`);
   } else {
     allSidebars = {};
   }
+
   Object.assign(allSidebars, versionFallback.sidebarData());
 
-  const order = {};
-
-  Object.keys(allSidebars).forEach(sidebar => {
-    const categories = allSidebars[sidebar];
-
-    let ids = [];
-    const categoryOrder = [];
-    Object.keys(categories).forEach(category => {
-      ids = ids.concat(categories[category]);
-      for (let i = 0; i < categories[category].length; i++) {
-        categoryOrder.push(category);
-      }
-    });
-
-    for (let i = 0; i < ids.length; i++) {
-      const id = ids[i];
-      let previous;
-      let next;
-      if (i > 0) previous = ids[i - 1];
-      if (i < ids.length - 1) next = ids[i + 1];
-      order[id] = {
-        previous,
-        next,
-        sidebar,
-        category: categoryOrder[i],
-      };
-    }
-  });
-  return order;
+  return createDocsIndex(allSidebars);
 }
 
 // process the metadata for a document found in either 'docs' or 'translated_docs'
@@ -347,7 +323,6 @@ function generateMetadataBlog() {
 
 module.exports = {
   getDocsPath,
-  readSidebar,
   processMetadata,
   generateMetadataDocs,
   generateMetadataBlog,
